refac general: maximum, minimum
