+++ ivy/functional/backends/mxnet/old/general.py
@@ -89,9 +89,6 @@ def dtype_bits(dtype_in):
 
 
 
-minimum = lambda x, y: _mx.nd.array(_mx.nd.minimum(_scalar_or_flat_array_to_scalar(x), _scalar_or_flat_array_to_scalar(y)))
-maximum = lambda x, y: _mx.nd.array(_mx.nd.maximum(_scalar_or_flat_array_to_scalar(x), _scalar_or_flat_array_to_scalar(y)))
-
