+++ ivy/functional/backends/numpy/old/general.py
@@ -80,8 +80,6 @@ def dtype_bits(dtype_in):
 
 
 
-minimum = np.minimum
-maximum = np.maximum
