+++ ivy/functional/backends/tensorflow/old/general.py
@@ -68,9 +68,6 @@ def dtype_bits(dtype_in):
 
 
 
-minimum = tf.minimum
-maximum = tf.maximum
-
