+++ ivy/functional/backends/torch/old/general.py
@@ -39,16 +39,6 @@ def dtype_bits(dtype_in):
 
 
 
-def minimum(x, y):
-    x_val = torch.tensor(x) if (isinstance(x, int) or isinstance(x, float)) else x
-    y_val = torch.tensor(y) if (isinstance(y, int) or isinstance(y, float)) else y
-    return torch.min(x_val, y_val)
-
-
-def maximum(x, y):
-    x_val = torch.tensor(x) if (isinstance(x, int) or isinstance(x, float)) else x
-    y_val = torch.tensor(y) if (isinstance(y, int) or isinstance(y, float)) else y
-    return torch.max(x_val, y_val)
 
 
 def cast(x, dtype_in: str):
