+++ ivy/functional/ivy/elementwise.py
@@ -1,6 +1,6 @@
 # global
 from typing import Union
-
+from numbers import Number
 # local
 import ivy
 from ivy.framework_handler import current_framework as _cur_framework
@@ -1134,3 +1134,30 @@ def erf(x: Union[ivy.Array, ivy.NativeArray])\
     :return: The Gauss error function of x.
     """
     return _cur_framework(x).erf(x)
+
+def minimum(x: Union[ivy.Array, ivy.NativeArray], y: Union[ivy.Array, ivy.NativeArray])\
+        -> Union[ivy.Array, ivy.NativeArray]:
+    """
+    Returns the min of x and y (i.e. x < y ? x : y) element-wise.
+
+    :param x: Input array containing elements to minimum threshold.
+    :type x: array
+    :param y: Tensor containing minimum values, must be broadcastable to x.
+    :type y: array
+    :return: An array with the elements of x, but clipped to not exceed the y values.
+    """
+    return _cur_framework(x).minimum(x, y)
+
+
+def maximum(x: Union[ivy.Array, ivy.NativeArray, Number], y: Union[ivy.Array, ivy.NativeArray, Number],
+            ) -> Union[ivy.Array, ivy.NativeArray]:
+    """
+    Returns the max of x and y (i.e. x > y ? x : y) element-wise.
+
+    :param x: Input array containing elements to maximum threshold.
+    :type x: array
+    :param y: Tensor containing maximum values, must be broadcastable to x.
+    :type y: array
+    :return: An array with the elements of x, but clipped to not be lower than the y values.
+    """
+    return _cur_framework(x).maximum(x, y)
